+++ subduction_websocket/tests/tokio_tests.rs
@@ -150,7 +150,7 @@ fn setup_client_subduction(
 ) -> (
     TestSubduction,
     TestHandler,
-    subduction_core::subduction::ListenerFuture<
+    subduction_core::subduction::listener_future::ListenerFuture<
         'static,
         Sendable,
         MemoryStorage,
@@ -178,7 +178,7 @@ fn setup_server_subduction(
 ) -> (
     ServerSubduction,
     ServerHandler,
-    subduction_core::subduction::ListenerFuture<
+    subduction_core::subduction::listener_future::ListenerFuture<
         'static,
         Sendable,
         MemoryStorage,
